Cloned from CL 300112829.
Refactor accept/connect handling.

This change refactors connect/accept handling so that the initial SYN/ SYN-ACK
is sent when connect is initiated or when the SYN is received without spinning
up the goroutine first. This should reduce the latency for accept/starting
connect.

This is a change to make the next step easier which is to eventually complete
connections in the dispatcher and not require a wake up to complete the
handshake.

Updates #231

PiperOrigin-RevId: 333885542